+++ src/tkc/emitter.h
@@ -303,6 +303,15 @@ ret_t emitter_forward(void* ctx, event_t* e);
 
 #define EMITTER(emitter) ((emitter_t*)(emitter))
 
+#define EMITTER_ENABLE(emitter)       \
+  if ((emitter) != NULL) {            \
+    emitter_enable(EMITTER(emitter)); \
+  }
+#define EMITTER_DISABLE(emitter)       \
+  if ((emitter) != NULL) {             \
+    emitter_disable(EMITTER(emitter)); \
+  }
+
 /*public for test*/
 ret_t emitter_remove_item(emitter_t* emitter, emitter_item_t* item);
 emitter_item_t* emitter_get_item(emitter_t* emitter, uint32_t index);